Float16 IOSurface slots use real CVPixelBuffer with kCVPixelFormatType_OneComponent16Half
install(): float16/manifest dtype 'float16' uses Arena::new() which calls
  tribunus_arena_alloc() creating a CVPixelBuffer with 16-bit half-float
  pixel format — actual IOSurface properties match MTLPixelFormat::R16Float.
  Other slots use Arena::new_bytes() generic surface (format provided by
  texture descriptor). Removed new_image() post-alloc patching approach.

Shader: half bitcast uses as_type<ushort>(v.r) (16-to-16 bits, then
  implicit widen to ulong for atomic add), not as_type<uint> (32-bit read
  from 16-bit value). R16Uint shader uses texture2d<ushort> matching
  16-bit texel width.

+++ compute-core/src/backend/metal_consumer.rs
@@ -263,7 +263,7 @@ kernel void checksum(texture2d<half, access::read> in  [[texture(0)]],
                      device atomic_ulong* out [[buffer(0)]],
                      uint tid [[thread_position_in_grid]]) {
     half4 v = in.read(uint2(tid, 0));
-    uint bits = as_type<uint>(v.r);
+    uint bits = as_type<ushort>(v.r);
     atomic_fetch_add_explicit(out, (ulong)bits, memory_order_relaxed);
 }
 "#
